refactor who-we-are page
