+++ src/pages/who-we-are/_OurValues.tsx
@@ -33,7 +33,6 @@ const StyledHeader = styled(Header)`
         max-width: 602px;
     }
 `
-
 const StyledFlex = styled(Flex)`
     @media (max-width: 930px) {
         flex-direction: column;
@@ -41,7 +40,6 @@ const StyledFlex = styled(Flex)`
         width: unset;
     }
 `
-
 const Card = styled(Flex)`
     max-width: 384px;
     padding: 32px 44px 32px 32px;
@@ -65,12 +63,12 @@ const OurValues = () => {
             </StyledHeader>
             <StyledFlex width="820px" wrap="wrap">
                 {our_values.map((card, index) => (
-                    <Card key={card.title} index={index} direction="column" ai="start" jc="start">
-                        <img src={card.icon} alt="" />
+                    <Card key={card.id} index={index} direction="column" ai="start" jc="start">
+                        <img src={card.icon} alt="icon" />
                         <Header as="h4" padding="24px 0 8px" size="32px" align="left" type="unset">
                             {card.title}
                         </Header>
-                        <Text size="16px">{card.text} </Text>
+                        <Text size="16px">{card.text}</Text>
                     </Card>
                 ))}
             </StyledFlex>
